fixing bugs and verify people.controller

+++ src/backend/controllers/people.controller.js
@@ -9,13 +9,17 @@ class PeopleController {
     createPerson = async (req, res) => {
         const data = req.body;
         if (!data || !data.firstName || !data.lastName || !data.dni || !data.cuil || !data.birthday || !data.about) { return res.json400("Missing information!(C)"); };
+        const verifyDNI = await this.verifyPersonDNI(data.dni);
+        if (verifyDNI) { return res.json400("Person alredy Exist!(C)"); };
         const person = await this.pService.createOne(data);
         return res.json201(person);
     };
 
-    getPersonByid = async (req, res) => {
+    getPersonById = async (req, res) => {
         const { pid } = req.params;
+        if((pid.length !== 24)) {return res.json400("Invalid Person ID!(C)"); };
         const person = await this.verifyPersonFun(pid);
+        if (person === false ) { return res.json400("Invalid Person ID!(C)"); };
         if (person === null) { return res.json404("Person Not Found!(C)"); };
         return res.json200(person);
     };
@@ -25,13 +29,14 @@ class PeopleController {
     };
 
     getAllPeople = async (req, res) => {
-        const people = await this.pService.getAllPeople();
+        const people = await this.pService.readAll();
         if (people.length === 0) { return res.json404("People Not Found!(C)"); };
         return res.json200(people);
     };
 
     updatePersonById = async (req, res) => {
         const { pid } = req.params;
+        if((pid.length !== 24)) {return res.json400("Invalid Person ID!(C)"); };
         const data = req.body;
         if (!data) { return res.json400("No Information to Update!(C)"); };
         const person = await this.verifyPersonFun(pid);
@@ -42,6 +47,7 @@ class PeopleController {
 
     deletePersonById = async (req, res) => {
         const { pid } = req.params;
+        if((pid.length !== 24)) {return res.json400("Invalid Person ID!(C)"); };
         const person = await this.verifyPersonFun(pid);
         if (person === null) { return res.json404("Person Not Found!(C)"); };
         const personDeleted = await this.pService.destroyById(pid);
@@ -49,10 +55,16 @@ class PeopleController {
     };
 
     verifyPersonFun = async (pid) => {
-        if(!isValidObjectId(pid)) { return res.json400("Invalid Person ID!(C)"); };
+        if (!isValidObjectId(pid))  { return false; };
         const verify = await this.pService.readById(pid);
         if (!verify) { return null; }
         else { return verify; };
+    };
+
+    verifyPersonDNI = async (dni) => {
+        const verify = await this.pService.readByFilter({ dni });
+        if (!verify) { return false; }
+        else { return true };
     };
 };
